GMMproject/requirements.txt
django==3.0.1
djangorestframework==3.11.0
django-rest-swagger==2.2.0
mysqlclient==1.4.6